Fixed Stack Delete Error
If there was only one slot open and player was standing on multiple mushrooms in the area that could be picked up, it would only pick up one of them and delete the rest. That doesnt happen anymore by checking at update per item for ItemCanBeSlotted in the Collectable script.