Request Letter Box
Basic operation to deposit in the inventory side and check against the random ones works.

Likely need to change the number of mushrooms that it can pull from. Advise to remove the day-by-day items and only ask for commons and non-daily items, otherwise you wont be able to submit until later in the game.