Updated Dialogue Typewriter
Roosevelt looked upon the original code and weeped...not really, but the solution to using TMP for typewriter and not having it attempt to dynamically size was to reveal by letter not UPDATE per letter.

This means dialogue objects can use styling instead of just getting pulled letter by lettter into the UI piece.