Day 1 Bug Fixes
Player was able to get around the cave lake. Added some blocking features and a harsh collision polygon to stop that. Updated mountain rock elements to y sort correctly and not use the tree fade script that just looked sorta weird on smaller elements. Updated the mountain pine trees to better y sort and collide at the base of the lower branches a bit more realistically.